+++ config/dynamo.js
